added comment panel below blog and corrected longs in Js classes

+++ src/de/threedimensions/blog/client/model/BlogEntryJs.java
@@ -13,7 +13,7 @@ public class BlogEntryJs extends JavaScriptObject {
     protected BlogEntryJs() {
     }
 
-    public final native Long getId() /*-{ return this.id; }-*/;
+    public final native int getId() /*-{ return this.id; }-*/;
 
     public final native String getHeading() /*-{ return this.heading; }-*/;
